Many a bug fix
Fixed numerous bugs:

One that caused a crash when trying to add "Stamp" to Circuit Court.

One that prevented Defendant perspective bills from updating the item list. (it's not fully implemented yet, just uses the plaintiff objects, but the behaviour was inconsistent)

Several items in the common items list weren't creating items when double-clicked, or were not disappearing from the list when they should have been.