Massive Update
This fixes a couple of major bugs, one with moving Witnesses Expenses to the next page, and another where changing a Counsel's fee wouldn;'t update the totals until the next time DOC_ORGANIZER was run. This was solved by making VAT items and lastPage_ProVAT update every step, and VAT will run DOC_ORGANIZER if something changed.

Also added the ability to change the VAT amount across the whole document. Ideally, we'll want to be able to change the VAT per item, with a warning similar to the Instructions Fees, but this will do for right now, since things can just be modified after exporting to Excel.

This change is saved and loaded in the cdf file, and as such, we now have a new docVer 2.2. We've also simplified the docVer checks, we were needlessly doing conversions before, now it just reads the 2.2. as a real value.