Very nearly there
Get it so that nothing else is clickable while the editor is active. Done.

Saving and loading. Done.

Have the items appear in the items list. Done.

Have double-clicking add the items as a "USER_ITEM" with the fields filled. Done.

Limit the length of entered descriptions to less than what is allowed normally. Done

Remove odd characters. Done