Many many things
Limited the text displayed in the editor adding "..." where a description length is too long for the field.

Mouse over an item displays the full text of the description. (note: changed how tooltips work so that they are multiline)

The bottom are always visible and clickable, regardless of window size.

The list stops displaying entries where the footer starts.

Fixed the behaviour of the view when moving items around the list.